moved SaveAndRestoreAuthenticationScope to model module and fixed user name for restored log entry
