+++ app/save-and-restore/model/pom.xml
@@ -17,6 +17,12 @@
 			<version>6.0.0-SNAPSHOT</version>
 		</dependency>
 
+      <dependency>
+          <groupId>org.phoebus</groupId>
+          <artifactId>core-security</artifactId>
+          <version>6.0.0-SNAPSHOT</version>
+      </dependency>
+
 	  <dependency>
 		  <groupId>org.phoebus</groupId>
 		  <artifactId>core-websocket-common</artifactId>
